// src/reward-job/config/reward-job-config.type.ts
export type RewardJobConfig = {
  stakingContractAddress?: string;
  tokenContractAddress?: string;
  vanaRpcUrl?: string;
  adminWalletPrivateKey?: string;
  adminWalletAddress?: string;
  duration7DaysRate?: string;
  duration30DaysRate?: string;
  duration60DaysRate?: string;
  duration90DaysRate?: string;
};